Initial commit of selected Boost 1.57.0 libraries.
Code taken from the official boost.org release,
Version 1.57.0, November 3rd, 2014 21:55 GMT.

The purpose of committing an embedded copy of Boost
is to reduce the risk of errors caused by version mismatch.
Only those Boost libraries that are actually used by this
and downstream projects are committed.

If the use of external dependencies is preferred,
remove dependencies/cpp from the include path.